+++ packages/terra-embedded-content-consumer/tests/wdio/embedded-content-provider-test-template-spec.js
@@ -1,6 +1,6 @@
 Terra.describeViewports('Embedded Content Provider Test Template', ['huge'], () => {
   it('has mounted and overridden the body height of 100%', () => {
-    browser.url('/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider');
+    browser.url('/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider');
     const bodyHeight = $('body').getSize('height');
     const viewportHeight = browser.getWindowSize('height');