+++ packages/terra-framework-docs/src/terra-dev-site/test/hookshot/common/HookshotContentTestTemplate.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import Hookshot from 'terra-hookshot';
+import styles from './HookshotTestDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
+
+const HookshotContent = props => (
+  <Hookshot.Content
+    {...props}
+  >
+    <div className={cx('hookshot-content-wrapper')}>Hookshot</div>
+  </Hookshot.Content>
+);
+
+export default HookshotContent;