+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel/SlidePanelLarge.test.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import SlidePanel, { SlidePanelPositions } from 'terra-slide-panel';
+import SlidePanelMainContent from './common/SlidePanelMainContent';
+import SlidePanelPanelContent from './common/SlidePanelPanelContent';
+
+const SlidePanelLarge = () => (
+  <SlidePanel
+    mainContent={<SlidePanelMainContent />}
+    panelContent={<SlidePanelPanelContent />}
+    panelBehavior="overlay"
+    panelPosition={SlidePanelPositions.END}
+    panelSize="large"
+    isOpen
+  />
+);
+
+export default SlidePanelLarge;