+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs to terra-framework-docs
+
 ## 4.41.0 - (June 8, 2021)
 
 * Changed