+++ packages/terra-framework-docs/CHANGELOG.md
@@ -20,6 +20,13 @@
   * Migrate docs from terra-tabs
   * Migrate docs from terra-slide-panel-manager
   * Migrate docs from terra-infinite-list
+  * Migrate docs from terra-hookshot
+  * Migrate docs from terra-disclosure-manager
+  * Migrate docs from terra-embedded-content-consumer
+  * Migrate docs from terra-slide-panel.
+  * Migrate docs from terra-slide-group.
+  * Migrate docs from terra-popup
+
 
 ## 1.7.0 - (March 8, 2022)